Add a dummy hw driver
This dummy driver must be used when no other low-level hardware
driver should be used for magic key detection.
Reason is, that the hw_init function seems to have a problem when
no driver is used at all. So with this dummy driver, it possible
to create the required one entry in the initcall list.

This can be removed when the assembly function in core/entry.S
is fixed.

Signed-off-by: Michael Heimpold <michael.heimpold@i2se.com>