linux_prep: do not pass initrd atag
For Duckbill 2 devices, we are not going to use initrd in favour
of initramfs. So it's not required to pass this ATAG from bootlets.

This also prevents a warning of the kernel:
INITRD: 0x40800000+0x02000000 overlaps in-use memory region - disabling initrd

When initrd support is needed, then revert this patch and adopt
the initrd load offset and size accordingly.

Signed-off-by: Michael Heimpold <michael.heimpold@i2se.com>